+++ app-server/cmd/main/main.go
@@ -2,13 +2,13 @@ package main
 
 import (
 	"app-server/internal/repository"
-	"app-server/internal/utilis"
+	"app-server/internal/utilities"
 )
 
 func main() {
 	database := repository.InitDB()
 	// repository.PerformMigration(database)
 
-	r := utilis.ServerRouter(database.GetDB())
+	r := utilities.ServerRouter(database.GetDB())
 	r.Run()
 }
