+++ app-server/internal/utilities/environment.go
@@ -1,4 +1,4 @@
-package utilis
+package utilities
 
 import "os"
