+++ app-server/internal/utilities/logger.go
@@ -1,4 +1,4 @@
-package utilis
+package utilities
 
 import (
 	"github.com/gin-gonic/gin"
